requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
Django==4.1.13
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djongo==1.3.6
dnspython==2.6.1
PyJWT==2.9.0
pymongo==4.8.0
redis==5.0.8
sqlparse==0.2.4
typing_extensions==4.12.2
